Audit storage size numbers were incorrect. Ported storage size startup reporting behavior from primary instance to audit instance.
